% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Season10}
\alias{Season10}
\title{Season10}
\format{A data frame with 19 rows and 9 variables:
\describe{
  \item{\code{Rank Tier}}{character The Competitive Rank Tier}
  \item{\code{Standard}}{numeric Competitive Standard 3vs3}
  \item{\code{Doubles}}{numeric Competitive Doubles 2vs2}
  \item{\code{Solo Duel}}{numeric Competitive Duel 1vs1}
  \item{\code{Solo Standard}}{numeric Competitive Solo Standard 3vs3}
  \item{\code{Rumble}}{numeric Competitive Rumble}
  \item{\code{Dropshot}}{numeric Competitive Dropshot}
  \item{\code{Hoops}}{numeric Competitive Hoops}
  \item{\code{Snow Day}}{numeric Competitive Snow Day}
}}
\source{
\url{https://www.reddit.com/r/RocketLeague/comments/bvac8t/season_10_rank_distribution/}
}
\usage{
Season10
}
\description{
Rank Distribution for Competitive Rocket League Season 10
}
\keyword{datasets}
