% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Season6}
\alias{Season6}
\title{Season6}
\format{A data frame with 19 rows and 5 variables:
\describe{
  \item{\code{Rank Tier}}{character The Competitive Rank Tier}
  \item{\code{Standard}}{numeric Competitive Standard 3vs3}
  \item{\code{Doubles}}{numeric Competitive Doubles 2vs2}
  \item{\code{Solo Duel}}{numeric Competitive Duel 1vs1}
  \item{\code{Solo Standard}}{numeric Competitive Solo Standard 3vs3}
}}
\source{
\url{https://www.reddit.com/r/RocketLeague/comments/7wgrm5/season_6_rank_distribution_data/}
}
\usage{
Season6
}
\description{
Rank Distribution for Competitive Rocket League Season 6
}
\keyword{datasets}
