% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Season7}
\alias{Season7}
\title{Season7}
\format{A data frame with 19 rows and 5 variables:
\describe{
  \item{\code{Rank Tier}}{character The Competitive Rank Tier}
  \item{\code{Standard}}{numeric Competitive Standard 3vs3}
  \item{\code{Doubles}}{numeric Competitive Doubles 2vs2}
  \item{\code{Solo Duel}}{numeric Competitive Duel 1vs1}
  \item{\code{Solo Standard}}{numeric Competitive Solo Standard 3vs3}
}}
\source{
\url{https://www.reddit.com/r/RocketLeague/comments/8omfoq/holy_dominus_its_the_season_7_rank_distribution/}
}
\usage{
Season7
}
\description{
Rank Distribution for Competitive Rocket League Season 7
}
\keyword{datasets}
